kyberjs: rename group module to curves and expose abstract classes
exposes kyber.Group, kyber.Point and kyber.Scalar and changes
kyber.group module to kyber.curves to avoid ambiguity
