+++ external/js/kyber/index.js
@@ -2,5 +2,10 @@
 
 const kyber = exports;
 
-kyber.group = require("./lib/group");
+kyber.curves = require("./lib/group");
 kyber.sign = require("./lib/sign");
+const abstractCls = require("./lib/index.js");
+
+kyber.Group = abstractCls.Group;
+kyber.Point = abstractCls.Point;
+kyber.Scalar = abstractCls.Scalar;
