+++ external/js/kyber/lib/group/nist/scalar.js
@@ -3,15 +3,15 @@
 const BN = require("bn.js");
 const crypto = require("crypto");
 const random = require("../../random");
-const group = require("../group.js");
+const group = require("../../index.js");
 
 /**
- * @module group/nist/scalar
+ * @module curves/nist/scalar
  */
 
 /**
  * Scalar
- * @param {module:group/nist/curve~Weirstrass} curve
+ * @param {module:curves/nist/curve~Weirstrass} curve
  * @param {BN.Red} red - BN.js Reduction context
  * @constructor
  */
@@ -30,7 +30,7 @@ class Scalar extends group.Scalar {
   /**
    * Equality test for two Scalars derived from the same Group
    *
-   * @param {module:group/nist/scalar~Scalar} s2 Scalar
+   * @param {module:curves/nist/scalar~Scalar} s2 Scalar
    * @return {boolean}
    */
   equal(s2) {
@@ -40,8 +40,8 @@ class Scalar extends group.Scalar {
   /**
    * Sets the receiver equal to another Scalar a
    *
-   * @param {module:group/nist/scalar~Scalar} a Scalar
-   * @return {module:group/nist/scalar~Scalar}
+   * @param {module:curves/nist/scalar~Scalar} a Scalar
+   * @return {module:curves/nist/scalar~Scalar}
    */
   set(a) {
     this.ref = a.ref;
@@ -51,7 +51,7 @@ class Scalar extends group.Scalar {
   /**
    * Returns a copy of the scalar
    *
-   * @return {module:group/nist/scalar~Scalar}
+   * @return {module:curves/nist/scalar~Scalar}
    */
   clone() {
     return new Scalar(this.ref.curve, this.ref.red).setBytes(
@@ -62,7 +62,7 @@ class Scalar extends group.Scalar {
   /**
    * Set to the additive identity (0)
    *
-   * @return {module:group/nist/scalar~Scalar}
+   * @return {module:curves/nist/scalar~Scalar}
    */
   zero() {
     this.ref.arr = new BN(0, 16).toRed(this.ref.red);
@@ -72,9 +72,9 @@ class Scalar extends group.Scalar {
   /**
    * Set to the modular sums of scalars s1 and s2
    *
-   * @param {module:group/nist/scalar~Scalar} s1 Scalar
-   * @param {module:group/nist/scalar~Scalar} s2 Scalar
-   * @return {module:group/nist/scalar~Scalar} s1 + s2
+   * @param {module:curves/nist/scalar~Scalar} s1 Scalar
+   * @param {module:curves/nist/scalar~Scalar} s2 Scalar
+   * @return {module:curves/nist/scalar~Scalar} s1 + s2
    */
   add(s1, s2) {
     this.ref.arr = s1.ref.arr.redAdd(s2.ref.arr);
@@ -84,9 +84,9 @@ class Scalar extends group.Scalar {
   /**
    * Set to the modular difference
    *
-   * @param {module:group/nist/scalar~Scalar} s1 Scalar
-   * @param {module:group/nist/scalar~Scalar} s2 Scalar
-   * @return {module:group/nist/scalar~Scalar} s1 - s2
+   * @param {module:curves/nist/scalar~Scalar} s1 Scalar
+   * @param {module:curves/nist/scalar~Scalar} s2 Scalar
+   * @return {module:curves/nist/scalar~Scalar} s1 - s2
    */
   sub(s1, s2) {
     this.ref.arr = s1.ref.arr.redSub(s2.ref.arr);
@@ -96,8 +96,8 @@ class Scalar extends group.Scalar {
   /**
    * Set to the modular negation of scalar a
    *
-   * @param {module:group/nist/scalar~Scalar} a Scalar
-   * @return {module:group/nist/scalar~Scalar}
+   * @param {module:curves/nist/scalar~Scalar} a Scalar
+   * @return {module:curves/nist/scalar~Scalar}
    */
   neg(a) {
     this.ref.arr = a.ref.arr.redNeg();
@@ -107,7 +107,7 @@ class Scalar extends group.Scalar {
   /**
    * Set to the multiplicative identity (1)
    *
-   * @return {module:group/nist/scalar~Scalar}
+   * @return {module:curves/nist/scalar~Scalar}
    */
   one() {
     this.ref.arr = new BN(1, 16).toRed(this.ref.red);
@@ -117,9 +117,9 @@ class Scalar extends group.Scalar {
   /**
    * Set to the modular products of scalars s1 and s2
    *
-   * @param {module:group/nist/scalar~Scalar} s1
-   * @param {module:group/nist/scalar~Scalar} s2
-   * @return {module:group/nist/scalar~Scalar}
+   * @param {module:curves/nist/scalar~Scalar} s1
+   * @param {module:curves/nist/scalar~Scalar} s2
+   * @return {module:curves/nist/scalar~Scalar}
    */
   mul(s1, s2) {
     this.ref.arr = s1.ref.arr.redMul(s2.ref.arr);
@@ -129,9 +129,9 @@ class Scalar extends group.Scalar {
   /**
    * Set to the modular division of scalar s1 by scalar s2
    *
-   * @param {module:group/nist/scalar~Scalar} s1
-   * @param {module:group/nist/scalar~Scalar} s2
-   * @return {module:group/nist/scalar~Scalar}
+   * @param {module:curves/nist/scalar~Scalar} s1
+   * @param {module:curves/nist/scalar~Scalar} s2
+   * @return {module:curves/nist/scalar~Scalar}
    */
   div(s1, s2) {
     this.ref.arr = s1.ref.arr.redMul(s2.ref.arr.redInvm());
@@ -141,8 +141,8 @@ class Scalar extends group.Scalar {
   /**
    * Set to the modular inverse of scalar a
    *
-   * @param {module:group/nist/scalar~Scalar} a
-   * @return {module:group/nist/scalar~Scalar}
+   * @param {module:curves/nist/scalar~Scalar} a
+   * @return {module:curves/nist/scalar~Scalar}
    */
   inv(a) {
     this.ref.arr = a.ref.arr.redInvm();
@@ -155,7 +155,7 @@ class Scalar extends group.Scalar {
    * @param {Uint8Array} b
    *
    * @throws {TypeError} when b is not Uint8Array
-   * @return {module:group/nist/scalar~Scalar}
+   * @return {module:curves/nist/scalar~Scalar}
    */
   setBytes(b) {
     if (b.constructor !== Uint8Array) {
@@ -185,7 +185,7 @@ class Scalar extends group.Scalar {
    * Set to a random scalar
    *
    * param {function} [callback] - to generate randomBytes of given length
-   * @return {module:group/nist/scalar~Scalar}
+   * @return {module:curves/nist/scalar~Scalar}
    */
   pick(callback) {
     callback = callback || crypto.randomBytes;
